+++ Descension/Assets/Prefab/UI/HUD/HotbarSlot.prefab.meta
