+++ Descension/Assets/Prefab/UI/Shop.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 69f20580b1726fb48992bb2ada580367
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
