+++ Descension/Assets/Prefab/UI/Shop/ShopMenuItemPrefab.prefab.meta
